TODO: Include & ViewStub
I want the ViewStub layouts to be interchangeable. Meaning if a user chooses a layout, then picks a different one, and decides to go back to the original pick, they can.

Right now, the include will show the layout picked in the newboard fragment, but you can not pick the same one twice. The user cannot change their mind and go back to a previous layout.